+++ examples/routes/jsx.tsx
@@ -1,4 +1,5 @@
 import { renderHtmlResponse } from "@http/html-stream";
+import { logTokens } from "@http/token-stream/transform/log-tokens";
 import { tokenize } from "@http/html-stream/transform/tokenize";
 import { tagHooks } from "@http/html-stream/transform/tag-hooks";
 import { prettify } from "@http/html-stream/hooks/prettify";
@@ -8,6 +9,7 @@ export function GET(req: Request, match: URLPatternResult) {
     <Page req={req} path={match.pathname.input} />,
     {
       transformers: [
+        logTokens(),
         tokenize(),
         tagHooks(...prettify()),
       ],
@@ -16,23 +18,31 @@ export function GET(req: Request, match: URLPatternResult) {
 }
 
 function Page({ req, path }: { req: Request; path: string }) {
+  const missing = Promise.resolve("https://unpkg.com/missing.css@1.2.0");
   return (
     <html>
+      <head>
+        <link rel="stylesheet" href={missing} />
+      </head>
       <body>
-        <h1>JSX Streaming Example</h1>
-        <p>You are here: {path}</p>
-        <h2>Headers</h2>
-        <table>
-          <thead>
-            <tr>
-              <th>Name</th>
-              <th>Value</th>
-            </tr>
-          </thead>
-          <tbody>
-            <Rows entries={req.headers.entries()} />
-          </tbody>
-        </table>
+        <header>
+          <h1>JSX Streaming Example</h1>
+          <p>You are here: {path}</p>
+        </header>
+        <main>
+          <h2>Headers</h2>
+          <table>
+            <thead>
+              <tr>
+                <th>Name</th>
+                <th>Value</th>
+              </tr>
+            </thead>
+            <tbody>
+              <Rows entries={req.headers.entries()} />
+            </tbody>
+          </table>
+        </main>
       </body>
     </html>
   );